Module smllr::vfs

Define a mock filesystem for more fine-grained control over unit tests

Structs

DeviceId
DeviceId is a wraper around a 'long' with several traits represents a device id.

Inode
Inode is wraper around a 'long' with several added traits (interface) which represents the inode of a file

RealFileSystem
Empty struct representing the Real Filesystem. Calls go through std::fs to access real files

TestFile
TestFile denotes a mockfile. Note that we are mocking the linux-style filesystem where many things are 'files', including directories, links, devices (/dev/sda might be familair).

TestFileSystem
Denotes a Mock Filesystem used instead of risking real data or dealing with the actual filesystem

TestMD
TestMD is the mock metadata struct.

Enums

FileType
Filetype is an enum of all types used for filesystem objects.

Traits

File
Define common interface for a real or mock file

MetaData
Define common interface for real of mock metadata object

VFS
The VFS [virtual file system] trait is the interface we require for the injection into the directectory walker.