Struct smllr::walker::DirWalker

pub struct DirWalker<T: VFS> { /* fields omitted */ }

Customizable object to traverse a series of directories, efficiently identifying files and omitting files in certain paths or that match certain patterns

Methods

impl<M, F, V> DirWalker<V> where
V: VFS<FileIter = F>,
F: File<MD = M>,
M: MetaData,

fn new<P: AsRef<Path>>(vfs: V, dirs: &[P]) -> DirWalker<V>

Create a new DirWalker from a list of directories

fn blacklist_folders(self, bl: Vec<&OsStr>) -> Self

Build up a DirWalker with a list of blacklisted folders

fn blacklist_patterns(self, bl: Vec<&str>) -> Self

Build up a DirWalker with a list of blacklisted path patterns

fn traverse_folder(&mut self, path: &Path)

Operate on a folder: iterate through its contents recursively

fn traverse_all(self) -> HashSet<PathBuf>

Collect all specified files into a set; this consumes the DirWalker